refactor(*): restructure file
