+++ pages/dashboard/index.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import DashboardComponent from '../../Components/Content/Dashboard';
+import Layout from '../../Components/Layout/Layout';
+
+function PageDashboard() {
+  return (
+    <Layout>
+      <DashboardComponent/>
+    </Layout>
+  );
+}
+
+export default PageDashboard
